Promjena Int u Float kod Namirnica.kt i NamirnicaRecepta.kt, implementirano azuriranje kod dodavanja kolicine namirnica u hladnjaku, obrisan nepotreban database paket #3

+++ Software/app/src/main/java/hr/foi/rampu/fridgium/entities/Namirnica.kt
@@ -5,7 +5,7 @@ import com.google.gson.annotations.SerializedName
 data class Namirnica(
     val id: Int,
     val naziv: String,
-    val kolicina_hladnjak: Int,
+    val kolicina_hladnjak: Float,
     @SerializedName("mjerna_jedinica_id") val mjernaJedinica: MjernaJedinica,
-    val kolicina_kupovina: Int
+    val kolicina_kupovina: Float
 )
