+++ Software/app/src/main/java/hr/foi/rampu/fridgium/entities/NamirnicaRecepta.kt
@@ -1,3 +1,3 @@
 package hr.foi.rampu.fridgium.entities
 
-data class NamirnicaRecepta (val recept_id : Int, val namirnica_id : Int, val kolicina: Int)
+data class NamirnicaRecepta (val recept_id : Int, val namirnica_id : Int, val kolicina: Float)
